requirements.txt
openai~=1.21.2
python-dotenv~=1.0.1